Refactor: moved restorer to separate package
